requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1


# NLP and AI
transformers==4.35.2

sentence-transformers==2.2.2
datasets==2.14.6
tokenizers==0.15.0
protobuf==6.31.1
sentencepiece==0.2.0

# Utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# File handling
Pillow==10.1.0
python-magic==0.4.27

# Environment and configuration
python-decouple==3.8
python-dotenv==1.0.0

# Async processing (optional)
celery==5.3.4
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
